feature/SA-42 Moves section created for accounts page
